+++ pkg/deployment/resources/inspector/acs_anonymous.go
@@ -35,7 +35,7 @@ func (p *arangoClusterSynchronizationsInspector) Anonymous(gvk schema.GroupVersi
 			if p.v1 == nil || p.v1.err != nil {
 				return nil, false
 			}
-			return &arangoClusterSynchronizationsInspectorAnonymousV1{i: p.v1}, true
+			return &arangoClusterSynchronizationsInspectorAnonymousV1{i: p.state}, true
 		}
 	}