+++ CHANGELOG.md
@@ -20,8 +20,13 @@
 - (Bugfix) Move Agency CommitIndex log message to Trace
 - (Feature) Force delete Pods which are stuck in init phase
 - (Bugfix) Do not tolerate False Bootstrap condition in UpToDate evaluation
+- (Improvement) Don't serialize and deprecate two DeploymentReplicationStatus fields
+- (Improvement) Improve error message when replication can't be configured
+- (Bugfix) Fix License handling in case of broken license secret
+- (Bugfix) Check ArangoSync availability without checking healthiness
+- (Improvement) Add Anonymous Inspector mods
+- (Improvement) Do not check checksums for DeploymentReplicationStatus.IncomingSynchronization field values
 - (Improvement) Multi-arch support for ID member
-
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress